massive changes

+++ src/pages/register.tsx
@@ -1,12 +1,12 @@
 import Title from "antd/es/typography/Title";
-import { Alert, Button, Form, Input } from "antd";
+import { Button, Form, Input, message } from "antd";
 import { useNavigate } from "react-router-dom";
 import { useRegisterMutation } from "@src/store/auth";
 import { useEffect, useState } from "react";
 import { useDispatch } from "react-redux";
-import { useGetUserByUsernameQuery } from "@src/store/user";
 import { Loading } from "@src/components/Loading";
 import { setCredentials } from "@src/store/authSlice";
+import { useGetUsernameExistQuery } from "@src/store/user";
 
 export const RegisterPage = () => {
 	const [form] = Form.useForm();
@@ -14,20 +14,27 @@ export const RegisterPage = () => {
 	const [register] = useRegisterMutation();
 	const dispatch = useDispatch();
 	const [username, setUsername] = useState("");
-	const {
-		data: user,
-		isLoading: userLoading,
-		error: userError,
-	} = useGetUserByUsernameQuery(username);
+	const { data: userExist, isLoading: userExistLoading } =
+		useGetUsernameExistQuery(username);
 	const onFinish = async () => {
 		const data = form.getFieldsValue();
 		setUsername(data.username);
 		const res = await register(data);
-		if (res) {
+		if (userExist) {
+			message.error("");
+		}
+		if (res && !userExist) {
 			localStorage.setItem("token", res.data.accessToken);
+			localStorage.setItem(
+				"userId",
+				JSON.stringify(res.data.userId)
+			);
 			navigate("/");
 			dispatch(
-				setCredentials({ user, token: res.data.accessToken })
+				setCredentials({
+					userId: res.data.userId,
+					token: res.data.accessToken,
+				})
 			);
 		}
 	};
@@ -36,17 +43,8 @@ export const RegisterPage = () => {
 		const token = localStorage.getItem("token");
 		if (token) navigate("/");
 	}, []);
-	if (userLoading) return <Loading />;
-	if (userError) {
-		console.error(userError);
-		return (
-			<Alert
-				message="Error fetching user"
-				type="error"
-				closable
-			/>
-		);
-	}
+	if (userExistLoading) return <Loading />;
+
 	return (
 		<div className="text-center flex items-center justify-center h-[100vh]">
 			<div>
